+++ Changes.md
@@ -6,6 +6,11 @@ Improvements
 
 - Catalogue : Improved performance when interacting with large Catalogues.
 
+Fixes
+-----
+
+- Catalogue : Removed incorrect error message on empty Catalogues
+
 1.1.1.0 (relative to 1.1.0.0)
 =======